refactor(AvailableTimePicker): extract EachDateBlock from AvailableTimePicker
